feat: 랭크 조회 기능 구현

+++ AD/relpl/domain/src/main/java/com/gdd/domain/model/rank/RankItem.kt
@@ -2,5 +2,5 @@ package com.gdd.domain.model.rank
 
 data class RankItem(
     val  nickname: String,
-    val distance: Double
+    val distance: Int
 )
